+++ csense-kotlin-common/src/main/kotlin/csense/kotlin/extensions/primitives/String.kt
@@ -385,3 +385,4 @@ inline fun String.fromHexStringToByteArray(): ShortArray? {
     return result
 }
 //endregion
+