+++ csense-kotlin-common/src/test/kotlin/csense/kotlin/extensions/collections/map/MutableMapKtTest.kt
@@ -41,13 +41,13 @@ class MutableMapKtTest {
 
     @Ignore
     @Test
-    fun setIfNotEmpty(){
+    fun setIfNotEmpty() {
 
     }
 
     @Ignore
     @Test
-    fun putSubList(){
+    fun putSubList() {
 
     }