+++ csense-kotlin-common/src/test/kotlin/csense/kotlin/logger/LKtTest.kt
@@ -2,7 +2,6 @@ package csense.kotlin.logger
 
 import csense.kotlin.*
 import csense.kotlin.extensions.collections.*
-import csense.kotlin.extensions.collections.list.*
 import csense.kotlin.test.assertions.*
 import kotlin.test.*