+++ csense-kotlin-jvm/src/main/kotlin/csense/kotlin/extensions/GeneralJvm.kt
@@ -3,7 +3,7 @@ package csense.kotlin.extensions
 /**
  * Like typeK except it converts to java class types instead.
  *
- * @see typeK
+ * see typeK
  * @return Class<T>
  */
 inline fun <reified T : Any> type(): Class<T> = T::class.java