+++ csense-kotlin-jvm/src/main/kotlin/csense/kotlin/logger/LExtensions.kt
@@ -8,7 +8,7 @@ import kotlin.reflect.*
 
 /**
  * Uses the simple class name for the tag, otherwise just as the regular debug method
- * @see L.debug
+ * see L.debug
  */
 inline fun LLogger.debug(kClass: KClass<*>, message: String, exception: Throwable? = null) {
     debug(kClass.java, message, exception)
@@ -16,7 +16,7 @@ inline fun LLogger.debug(kClass: KClass<*>, message: String, exception: Throwabl
 
 /**
  * Uses the simple class name for the tag, otherwise just as the regular debug method
- * @see L.warning
+ * see L.warning
  */
 inline fun LLogger.warning(kClass: KClass<*>, message: String, throwable: Throwable? = null) {
     warning(kClass.java, message, throwable)
@@ -24,7 +24,7 @@ inline fun LLogger.warning(kClass: KClass<*>, message: String, throwable: Throwa
 
 /**
  * Uses the simple class name for the tag, otherwise just as the regular debug method
- * @see L.error
+ * see L.error
  */
 inline fun LLogger.error(kClass: KClass<*>, message: String, exception: Throwable? = null) {
     error(kClass.java, message, exception)
@@ -32,7 +32,7 @@ inline fun LLogger.error(kClass: KClass<*>, message: String, exception: Throwabl
 
 /**
  * Uses the simple class name for the tag, otherwise just as the regular debug method
- * @see L.logProd
+ * see L.logProd
  */
 inline fun LLogger.logProd(kClass: KClass<*>, message: String, throwable: Throwable? = null) {
     logProd(kClass.java, message, throwable)
@@ -41,7 +41,7 @@ inline fun LLogger.logProd(kClass: KClass<*>, message: String, throwable: Throwa
 
 /**
  * Uses the simple class name for the tag, otherwise just as the regular debug method
- * @see L.debug
+ * see L.debug
  */
 inline fun LLogger.debug(cls: Class<*>, message: String, exception: Throwable? = null) {
     debug(cls.simpleName, message, exception)
@@ -49,7 +49,7 @@ inline fun LLogger.debug(cls: Class<*>, message: String, exception: Throwable? =
 
 /**
  * Uses the simple class name for the tag, otherwise just as the regular debug method
- * @see L.warning
+ * see L.warning
  */
 inline fun LLogger.warning(cls: Class<*>, message: String, throwable: Throwable? = null) {
     warning(cls.simpleName, message, throwable)
@@ -57,7 +57,7 @@ inline fun LLogger.warning(cls: Class<*>, message: String, throwable: Throwable?
 
 /**
  * Uses the simple class name for the tag, otherwise just as the regular debug method
- * @see L.error
+ * see L.error
  */
 inline fun LLogger.error(cls: Class<*>, message: String, exception: Throwable? = null) {
     error(cls.simpleName, message, exception)
@@ -65,7 +65,7 @@ inline fun LLogger.error(cls: Class<*>, message: String, exception: Throwable? =
 
 /**
  * Uses the simple class name for the tag, otherwise just as the regular debug method
- * @see L.logProd
+ * see L.logProd
  */
 inline fun LLogger.logProd(cls: Class<*>, message: String, throwable: Throwable? = null) {
     logProd(cls.simpleName, message, throwable)